+++ src/main/java/com/robertx22/age_of_exile/repair_kits/RepairKitItem.java
@@ -88,7 +88,7 @@ public class RepairKitItem extends Item implements IAutoLocName, IAutoModel, ISh
         if (tier == 0) {
             return shaped(this)
                 .input('v', recipeItem)
-                .input('t', ModRegistry.MISC_ITEMS.MAGIC_ESSENCE)
+                .input('t', ModRegistry.MISC_ITEMS.T1_DUST)
                 .pattern(" v ")
                 .pattern("vtv")
                 .pattern(" v ")