+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/blocks/slots/FuelSlot.java
@@ -1,6 +1,5 @@
 package com.robertx22.age_of_exile.vanilla_mc.blocks.slots;
 
-import com.robertx22.age_of_exile.mmorpg.ModRegistry;
 import net.minecraft.inventory.Inventory;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -24,12 +23,9 @@ public class FuelSlot extends Slot {
             {
                 put(Items.DIAMOND, 500);
                 put(Items.GOLD_INGOT, 100);
-                put(Items.IRON_INGOT, 30);
-                put(Items.EMERALD, 200);
+                put(Items.IRON_INGOT, 50);
+                put(Items.EMERALD, 100);
                 put(Items.REDSTONE, 4);
-
-                put(ModRegistry.MISC_ITEMS.MAGIC_ESSENCE, 50);
-                put(ModRegistry.MISC_ITEMS.RARE_MAGIC_ESSENCE, 250);
             }
         }
     };